+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 
+## [0.0.3] - New Features Added!!
+
+- *onlyIcon* property added :
+    if only icon is needed to be shown in splash-screen .
+
 ## [0.0.2] - New Features Added!!
 
 - Custom Page Transition's Added .